feat(game-template): move match-generator-base to game-template to support loading config from assets
