+++ game-utils/game-template/match-generator/core/build.gradle
