+++ game-utils/game-template/match-generator/core/src/com/planet_lia/match_generator/libs/Timer.java
@@ -1,0 +1,8 @@
+package com.planet_lia.match_generator.libs;
+
+/**
+ * Timer holds time in seconds.
+ */
+public class Timer {
+    public float time = 0;
+}
