+++ game-utils/game-template/match-generator/core/src/com/planet_lia/match_generator/logic/Args.java
@@ -1,0 +1,16 @@
+package com.planet_lia.match_generator.logic;
+
+
+import com.planet_lia.match_generator.libs.DefaultArgs;
+
+public class Args extends DefaultArgs {
+    // Here you can put additional arguments for your own game
+    // that are not already specified in DefaultArgs class.
+    // Note to import com.beust.jcommander.Parameter.
+    // eg.
+    // ...
+    // @Parameter(names = { "--map-seed", "-m" },
+    //            description = "Seed for random number generator for map generation")
+    // public int mapSeed = 0;
+}
+
