+++ game-utils/game-template/match-generator/settings.gradle
