--- game-utils/match-generator-base/build.gradle
@@ -1,72 +1,0 @@
-buildscript {
-    ext.kotlinVersion = '1.3.41'
-
-    repositories {
-        mavenLocal()
-        mavenCentral()
-        maven { url "https://plugins.gradle.org/m2/" }
-        maven { url "https://oss.sonatype.org/content/repositories/snapshots/" }
-        jcenter()
-        google()
-    }
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
-    }
-}
-
-allprojects {
-    apply plugin: "idea"
-
-    version = '1.0'
-    ext {
-        appName = "match-generator-base"
-        gdxVersion = '1.9.10'
-        roboVMVersion = '2.3.7'
-        box2DLightsVersion = '1.4'
-        ashleyVersion = '1.7.0'
-        aiVersion = '1.8.0'
-        jCommanderVersion = '1.71'
-        javaWebSocketVersion = '1.4.0'
-    }
-
-    repositories {
-        mavenLocal()
-        mavenCentral()
-        jcenter()
-        google()
-        maven { url "https://oss.sonatype.org/content/repositories/snapshots/" }
-        maven { url "https://oss.sonatype.org/content/repositories/releases/" }
-    }
-}
-
-project(":desktop") {
-    apply plugin: "kotlin"
-
-    dependencies {
-        implementation project(":core")
-        compile "com.badlogicgames.gdx:gdx-backend-lwjgl3:$gdxVersion"
-        api "com.badlogicgames.gdx:gdx-platform:$gdxVersion:natives-desktop"
-        api "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
-        compile "com.badlogicgames.gdx:gdx-backend-headless:$gdxVersion"
-    }
-}
-
-project(":core") {
-    apply plugin: "kotlin"
-
-    sourceSets.test.java.srcDirs = ["test/"]
-
-    dependencies {
-        api "com.badlogicgames.gdx:gdx:$gdxVersion"
-        api "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
-        compile "com.beust:jcommander:$jCommanderVersion"
-        compile "org.java-websocket:Java-WebSocket:$javaWebSocketVersion"
-
-        testImplementation 'org.junit.jupiter:junit-jupiter-api:5.3.1'
-        testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.3.1'
-    }
-
-    test {
-        useJUnitPlatform()
-    }
-}
