--- game-utils/match-generator-base/core/src/com/planet_lia/match_generator_base/libs/Timer.java
@@ -1,8 +1,0 @@
-package com.planet_lia.match_generator_base.libs;
-
-/**
- * Timer holds time in seconds.
- */
-public class Timer {
-    public float time = 0;
-}
