--- game-utils/match-generator-base/core/src/com/planet_lia/match_generator_base/logic/Args.java
@@ -1,16 +1,0 @@
-package com.planet_lia.match_generator_base.logic;
-
-
-import com.planet_lia.match_generator_base.libs.DefaultArgs;
-
-public class Args extends DefaultArgs {
-    // Here you can put additional arguments for your own game
-    // that are not already specified in DefaultArgs class.
-    // Note to import com.beust.jcommander.Parameter.
-    // eg.
-    // ...
-    // @Parameter(names = { "--map-seed", "-m" },
-    //            description = "Seed for random number generator for map generation")
-    // public int mapSeed = 0;
-}
-
